grades_management_backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
bcrypt
pandas
